[App] Added React Navigation
