+++ Browse.js
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native';
+
+function BrowseScreen () {
+    return (
+        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+         <Text>Browse Screen</Text>
+        </View>
+    );
+    
+}
+
+export default BrowseScreen;
