+++ Options.js
@@ -1,0 +1,9 @@
+import { Text, View, Pressable, TouchableOpacity } from 'react-native';
+
+const OptionsScreen = ({ navigation }) => {
+    return (
+      <Text>Test Options Screen</Text>
+    );
+  }
+
+export default OptionsScreen;
